requirements.txt
streamlit==1.29.0
openai==1.6.1
python-dotenv==1.0.0
langchain==0.0.352
chromadb==0.4.18
pydantic==2.5.2
pydantic-settings==2.1.0
azure-storage-blob==12.19.0
numpy>=1.22.5
pypdf==3.17.1
